Handle  the instance migrations from the evac_status job
This is for the client-side part. The evac-node_Status job gets a full
JSON with the instance mirgation status and applied the changes for each
instance.

The "progress" part of the instance in the returned JSON now containes
the current task status of each migrate job.

Realigned the template, so this diff isn't really readable.